requirements-whisper.txt
aiogram
python-telegram-bot
SpeechRecognition
pydub
yt-dlp
python-dotenv
aiohttp
openai-whisper
deepmultilingualpunctuation
